feat(config): use HOCON format instead of YAML
The motivation to change the config format come from a limitation of Configurate about YAML format. It is not possible to serialize a YAML content with comments whereas this is a requirement for the project.
More details can be found here about Configurate and YAML module: https://github.com/SpongePowered/Configurate/pull/175