test: rework and improve integration tests
The ordering dependency between tests was not an ideal since it broke isolation and independence of tests.

Another limitation of this approach was the incompleteness of tests: only SQLite or MySQL were used depending on cases whereas it would be preferable to run all integration tests with all currently supported DBMS (SQLite, MySQL and MariaDB).

A major action taken here was to setup "base" abstract test classes containing all the tests to execute without depending on any DBMS. Then, this base class had just to be extended by each DBMS, add specific configurations if required and that's it. A big advantage of this approach is the possibility to run SQLite tests as unit tests one with Surefire which permits to have an earlier and quicker feedback whereas MySQL and MariaDB will be executed only with Failsafe because of environment setup prerequisites. In summary, the likelihood of having an issue with MySQL/MariaDB tests is low because of SQLite ones being executed first even if still possible.

It is important to note that for PatchPlaceBreakApi and PatchPlaceBreakBukkitAdapterApi, tests are duplicated with small differences between them. This is the decision cost of decoupling the core modules from Bukkit. However, since code dealing with Bukkit is clearly not easily testable and even not testable depending on cases, the costs are hugely under the benefits of this approach. So, the duplication is totally acceptable here even if not ideal.

Another important notice to have is about the random generation of BlockLocation for the Patch API and of Block for the Bukkit adapted one. This is required for MySQL and MariaDB in order to prevent collisions between tests. This has been done for performances purposes since this is not realistic to delete and recreate Docker containers between each test.

Small additional improvements:
* Use of @TempDir from JUnit to simplify temporary directory management
* Use of Arguments.arguments() and Named.named() factory methods instead of the "of()" ones in order to do static import, shorten a bit the code and make it easier to read

Special note:
* Add of "threeten-extra" dependency to get access to the MutableClock class which permits to have full control over the time in tests (especially useful to test ephemeral tags)