ci(github): no longer bypass tag protection at releasing time (#480)
A new ruleset named "Release tags" has been created instead of the initially created tag protection. This change allows tag creation without protection, which is acceptable since in the end that's the release creation which really matter. Since the tag creation action is no longer protected, we can get rid of the GitHub PAT "PUSH_TOKEN", leading in the end to a better security.