ci(github): merge release workflows
The initial motivation was to fix the following error: `${{ github.repository }}/.github/workflows/release-sign.yml@refs/heads/main is not allowed to be used in Djaytan/mc-jobs-reborn-patch-place-break`.

The easy workaround was to replace `${{ github.repository }}` by `Djaytan/mc-jobs-reborn-patch-place-break`, but it implies losing portability since the workflow would require modifications whenever it would be copied in another repository (as it is planned soon).

In order to avoid that, the usage of two workflows for releasing a new version as been questioned. The main initial reason for this split was because of the need to support the `workflow_dispatch` at signing time. This was required in order to sign old releases. Since it has now been done, supporting this type of events is no longer really justified. Thus, dropping the support of this event type provide a new opportunity: the possibility to merge the two release workflows together which end up simplifying the setup and fixing the issue while preserving the workflow portability across GitHub repositories. Furthermore, if one day a manual signing would be required, then we could simply execute some commands manually and upload the signature files in the associated GitHub release.

For the implementation, two possibilities were available: merging the signing logic right into the already existing release job, or creating a second job while depending on the initial one. In the end it has been decided to opt for the second option for security and clarity purposes. Taking care of having a clean workflow description is important since the signing logic is a bit complex, hence the fact it's better to separate this specific logic from the rest. While it could have been done through comments for highlighting the existence of two big steps in the job, the security best practices advice us to follow the principal of least privileges. Typically, the signing logic doesn't require all the permissions already provided to the main release logic (i.e. write accesses to `pull-requests` and `issues`) while requiring other ones not yet granted (i.e. generating an OIDC ID token by having write accesses for `id-token`). A same reflexion apply for the firewall. Thus, the best way in our context to follow the principal of least privileges is to split the workflow in two jobs. Furthermore, it doesn't require extra comments since having two separated jobs make it evident enough that the workflow is made of two major steps. A big argument against multi-jobs workflow in this context would be the performances cost implied, but in fact this is not a concern for a release since in any case it would take only few minutes to execute, which is reasonable for such a valuable workflow (manual releasing flow would require lots of additional time while being less safe & secure).

Based on the need to define a multi-job workflow, the following documentations have been followed in order to share a value between jobs in a workflow (i.e. the release tag):
* https://docs.github.com/en/actions/learn-github-actions/variables#passing-values-between-steps-and-jobs-in-a-workflow
* https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idneeds
* https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idoutputs