ci(github): simplify signature verification logic (#530)
This change is useful for an incoming one which consist on extracting the signature and verification logic in a script file. The final goal is to make local testing easier and improve the overall readability.

Here, redundant verifications are removed. Checking the workflow, the branch ref and the repository are already covered by the value of the `CERTIFICATE_IDENTITY` environment variable. The value is always expected to be `https://github.com/Djaytan/mc-jobs-reborn-patch-place-break/.github/workflows/release.yml@refs/heads/main`.

Realistically, we don't care about verifying which event triggered the signing process. The same apply for the commit SHA verification. Based on that, removing these specific verifications for simplifications purposes is better (i.e. the `--certificate-github-workflow-trigger` and `--certificate-github-workflow-sha` args respectively).