fix(release): resolve permission denied issue when executing script
It seems the `generate-plugin.sh` file is stored in Git without execution permission, which lead to the following error when trying to release a new version through the dedicated GitHub Actions workflow:

```
/bin/sh: 1: ./scripts/generate-plugin.sh: Permission denied
```

Trying to execute the command `chmod +x generate-plugin.sh` doesn't imply any change. Then I found the following article which provided me the right solution to fix the issue: https://www.designcise.com/web/tutorial/how-to-fix-github-action-permission-denied-error-when-executing-a-file.

The solution on Windows was to execute the following command instead:

```
git update-index --chmod=+x generate-plugin.sh
```