ci(release): install Java 17 before releasing a new version (#576)
When executing the `generate-plugin.sh` script through the `semantic-release` CLI, the following error is encountered:

```
Caused by: org.codehaus.plexus.compiler.CompilerException: error: release version 17 not supported
```

And it makes sense since the Java version has not been setup, meaning that the fix is trivial.

This time, when setting up the `actions/setup-java` action, no cache is leveraged since it's useless and thus will use our cache quota for nothing. Only the version is set and that's it.