fix(core): fail on overflow/underflow (#589)
The possibility to encounter weird behaviors because of overflow/underflow was known. That being said, this was not a priority to fix since the likelihood of such exploitation is potentially impossible.

This change aims to fail when moving a tag from one block to another at worlds' limits in a such way that an overflow or underflow occurs.