ci: allow 'chore' Conv. Commit type for dependencies updates & others (#595)
This change is motivated by the fact the `:semanticCommitType(build)` doesn't work as intended: PRs with commits prefixed by `chore` are still created. However, the initial decision is now questioned.

The `chore` is a well-known conventional commit type and is especially used for some dependencies updates and when changing specific files like the `.gitignore` one.

Furthermore, the `config:recommended` Renovate's preset relies on this type by default for implementing a "smart" logic (in short, it chooses the `fix` type for non-tests Maven dependencies, `chore` otherwise).
=> https://docs.renovatebot.com/semantic-commits/

Based on that, reintegrating the `chore` type seems to be the way to go for sharing a standard setup with the OSS community and wider.