build(packaging): explicitly define the "test" scope for dependencies
Providing an explicit scope to the dependencies defined in the `dependencyManagement` section doesn't affect the Maven execution. However, it becomes a lot easier for Renovate to identify the test dependencies. Retrieving accurate information regarding the dependency type allows him to infer the Conventional Commit type to apply:

* If the dependency is a test one, then the prefix `chore` apply.
* If that's a dependency used at compile and/or runtime by the application, then the prefix `fix` apply.
* If that's a Maven plugin dependency, then the prefix `chore` apply.
* If the dependency is not defined in Maven (e.g. the ones defined in the `package.json` file for CI dependencies), then the prefix `chore` apply.

=> https://docs.renovatebot.com/semantic-commits/

When taking a look to the Renovate's Dependency Dashboard of this repository (https://github.com/Djaytan/mc-jobs-reborn-patch-place-break/issues/6), we can see some mistakes:
* `equalsverifier` is a test dependency, but its prefix is `fix`
* Same for `testcontainers`

Hence this change for solving these two observable issues.

Since it's not possible to infer whether an `import` Maven scope is for test or production, it has been decided to abort usages of this scope for the test dependencies (i.e. JUnit, Mockito and Testcontainers). That's not ideal, but it's acceptable for improving the release note accuracy and correctness.